refactor: cleaning up
